Receipt Service and Controller
Receipt Service and Controller

+++ src/EnixerPos.Api.ViewModels/Sale/OrderItemModel.cs
@@ -6,7 +6,7 @@ namespace EnixerPos.Api.ViewModels.Sale
     {
         public int ItemName { get; set; }
         public decimal ItemPrice { get; set; }
-        public int ItemDiscount { get; set; }
+        public decimal ItemDiscount { get; set; }
         public bool IsDiscountPercentage { get; set; }
         public int Quantity { get; set; }
         public string OptionName { get; set; }
